\name{Euclidean1-package}
\alias{Euclidean1-package}
\alias{Euclidean1}
\docType{package}
\title{
\packageTitle{Euclidean1}
}
\description{
\packageDescription{Euclidean1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Euclidean1}
\packageIndices{Euclidean1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Euclidean1}

Maintainer: \packageMaintainer{Euclidean1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
